Extract delay into separate module
